+++ PanelDesign/src/edu/stanford/facs/panel/StainSet.java
@@ -3,7 +3,7 @@ package edu.stanford.facs.panel;
 public abstract class StainSet
 implements Comparable<StainSet>
 {
-	public final double index;
+	public final float index;
 	
 	public abstract int size ();
 	public abstract Marker marker (int i);
@@ -14,6 +14,8 @@ implements Comparable<StainSet>
 	@Override
 	public int compareTo (StainSet that)
 	{
+		// stains with higher staining index come first
+		// not consistent with equals since different stains may have the same index
 		if (this.index > that.index)
 			return -1;
 		if (this.index < that.index)
@@ -23,6 +25,6 @@ implements Comparable<StainSet>
 
 	protected StainSet (double index)
 	{
-		this.index = index;
+		this.index = (float) index;
 	}
 }
